server/requirements.txt
kafka_python==2.0.2
neo4j==5.25.0
psycopg2==2.9.5
